projects: Move Access enum into its own file
To structure things a bit more
